Add icon facility to Projectiles
Projectiles now have an optional class variable, icon, that can be set to an
image file. When the Projectile appears as a power-up, the icon is used if
available. If no icon is given, an image of the Projectile is used as before.
It is assumed that an icon image contains any necessary decorations--a
decorative box is NOT drawn around a Projecile icon.